# openapi/ontology.yaml
openapi: 3.1.0
info:
  title: Ontology OpenAPI
  description: Ontology OpenAPI Specification
  version: 0.0.1
servers:
  - url: http://localhost:4242
    description: Local development server
...
paths:
  "/v1/ontologies":
    get:
      tags:
        - Ontology
      summary: List Ontologies
      operationId: listOntologies
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
      responses:
        "200":
          description: Visible Ontologies
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/OntologyPage"
    post:
      tags:
        - Ontology
      summary: Create an Ontology
      operationId: createOntology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateOntologyReq"
      responses:
        "201":
          description: Created Ontology
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Ontology"
...
  /v1/ontologies/{ontologyId}/connections:
    get:
      tags:
        - DataConnection
      summary: List Connections
      operationId: listConnections
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: owner
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List Connections response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSourcePage'
    post:
      tags:
        - DataConnection
      summary: Create Connection
      operationId: createConnection
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequest'
        required: true
      responses:
        '200':
          description: Create Connection response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSource'
    parameters:
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/connections/{id}:
    get:
      tags:
        - DataConnection
      summary: Get Connection
      operationId: getConnection
      responses:
        '200':
          description: Get Connection response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSource'
    patch:
      tags:
        - DataConnection
      summary: Update Connection
      operationId: updateConnection
      parameters:
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRequest'
        required: true
      responses:
        '200':
          description: Update Connection response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSource'
    delete:
      tags:
        - DataConnection
      summary: Delete Connection
      operationId: deleteConnection
      responses:
        '200':
          description: Delete Connection response
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/connections/{id}/assets:
    get:
      tags:
        - DataConnection
      summary: List Connection Assets
      operationId: listConnectionAssets
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 50
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List Connection Assets response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPage'
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
...
  /v1/ontologies/{ontologyId}/connections/{id}/assets/{assetId}/infer-schema:
    post:
      tags:
        - DataConnection
      summary: Infer Schema
      operationId: inferSchema
      responses:
        '200':
          description: Infer Schema response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryRun'
    parameters:
      - name: assetId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/connections/{id}/assets/{assetId}/preview:
    post:
      tags:
        - DataConnection
      summary: Preview Connection Asset
      operationId: previewConnectionAsset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreviewRequest'
      responses:
        '200':
          description: Preview Connection Asset response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPreview'
    parameters:
      - name: assetId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/connections/{id}/assets/{assetId}/usage:
    get:
      tags:
        - DataConnection
      summary: Get Connection Asset Usage
      operationId: getConnectionAssetUsage
      responses:
        '200':
          description: Get Connection Asset Usage response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetUsage'
    parameters:
      - name: assetId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/connections/{id}/disable:
    post:
      tags:
        - DataConnection
      summary: Disable
      operationId: disable
      responses:
        '200':
          description: Disable response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSource'
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/connections/{id}/discover:
    post:
      tags:
        - DataConnection
      summary: Discover
      operationId: discover
      responses:
        '200':
          description: Discover response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryRun'
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
...
  /v1/ontologies/{ontologyId}/connections/{id}/overview:
    get:
      tags:
        - DataConnection
      summary: Get Connection Overview
      operationId: getConnectionOverview
      responses:
        '200':
          description: Get Connection Overview response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Overview'
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/connections/{id}/pipelines:
    get:
      tags:
        - DataConnection
      summary: List Connection Pipelines
      operationId: listConnectionPipelines
      responses:
        '200':
          description: List Connection Pipelines response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PipelineSummary'
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
...
  /v1/ontologies/{ontologyId}/connections/{id}/runs:
    get:
      tags:
        - DataConnection
      summary: List Connection Runs
      operationId: listConnectionRuns
      responses:
        '200':
          description: List Connection Runs response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PipelineRunSummary'
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
...
  /v1/ontologies/{ontologyId}/connections/local-csv:
    post:
      tags:
        - DataConnection
      summary: Import Local Csv
      operationId: importLocalCsv
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: description
          in: query
          required: false
          schema:
            type: string
        - name: tags
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - files
      responses:
        '200':
          description: Import Local Csv response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSource'
    parameters:
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/credentials:
    get:
      tags:
        - DataConnection
      summary: List Credentials
      operationId: listCredentials
      parameters:
        - name: usable
          in: query
          required: false
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: List Credentials response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CredentialSummary'
    parameters:
      - $ref: '#/components/parameters/OntologyId'
...
components:
  schemas:
    CreateOntologyReq:
      type: object
      additionalProperties: false
      required:
        - name
        - id
      properties:
        id:
          type: string
          maxLength: 32
        name:
          type: string
          minLength: 1
          maxLength: 32
        description:
          type: string
          maxLength: 1024
...
    OntologyPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/Ontology"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
    Ontology:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - createdAt
        - updatedAt
...
    DataSourcePage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DataSource'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
        counts:
          type: object
          additionalProperties:
            type: integer
            format: int32
        filters:
          type: object
          additionalProperties: {}
    CreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
            - S3_CSV
            - MYSQL
            - POSTGRESQL
            - KAFKA
            - EXTERNAL_PULSAR
        tags:
          type: array
          items:
            type: string
        config:
          type: object
          additionalProperties: {}
        credential:
          $ref: '#/components/schemas/CredentialInput'
        testToken:
          type: string
        ownerId:
          type: string
        ownerName:
          type: string
    DataSource:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
            - S3_CSV
            - MYSQL
            - POSTGRESQL
            - KAFKA
            - EXTERNAL_PULSAR
        status:
          type: string
          enum:
            - UNTESTED
            - TESTING
            - HEALTHY
            - HEALTHY_RESTRICTED
            - ERROR
            - DISABLED
        tags:
          type: array
          items:
            type: string
        config:
          type: object
          additionalProperties: {}
        credential:
          $ref: '#/components/schemas/CredentialSummary'
        syncStatus:
          type: string
          enum:
            - NO_TASKS
            - IDLE
            - RUNNING
            - STREAMING
            - PARTIAL_FAILURE
            - ALL_FAILURE
        assetCount:
          type: integer
          format: int32
        lastError:
          $ref: '#/components/schemas/Diagnostic'
        pipelineReferenceCount:
          type: integer
          format: int32
        activeRunCount:
          type: integer
          format: int32
        ownerId:
          type: string
        ownerName:
          type: string
        lastCheckedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        config:
          type: object
          additionalProperties: {}
        ownerId:
          type: string
        ownerName:
          type: string
    AssetPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DataSourceAsset'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    DataSourceAsset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        fullPath:
          type: string
        parentPath:
          type: string
        assetType:
          type: string
        schemaStatus:
          type: string
        schemaHash:
          type: string
        schemaVersion:
          type: integer
          format: int32
        fieldCount:
          type: integer
          format: int32
        sizeBytes:
          type: integer
          format: int64
        estimatedRows:
          type: integer
          format: int64
        partitionCount:
          type: integer
          format: int32
        permissionStatus:
          type: string
        usedByPipeline:
          type: boolean
        fields:
          type: array
          items:
            $ref: '#/components/schemas/AssetField'
        discoveredAt:
          type: string
          format: date-time
    DiscoveryRun:
      type: object
      properties:
        taskId:
          type: string
          format: uuid
        status:
          type: string
        discoveredCount:
          type: integer
          format: int32
        diagnostic:
          $ref: '#/components/schemas/Diagnostic'
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    PreviewRequest:
      type: object
      properties:
        nodeId:
          type: string
        limit:
          type: integer
          format: int32
    AssetPreview:
      type: object
      properties:
        columns:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: object
            additionalProperties: {}
        truncated:
          type: boolean
        maxBytes:
          type: integer
          format: int32
    AssetUsage:
      type: object
      properties:
        pipelines:
          type: array
          items:
            $ref: '#/components/schemas/PipelineSummary'
        restrictedReferences:
          type: integer
          format: int32
    Overview:
      type: object
      properties:
        connection:
          $ref: '#/components/schemas/DataSource'
        health:
          type: array
          items:
            $ref: '#/components/schemas/TestStage'
        assetSummary:
          type: object
          additionalProperties:
            type: integer
            format: int32
        pipelineSummary:
          type: object
          additionalProperties:
            type: integer
            format: int32
        recentRuns:
          type: array
          items:
            $ref: '#/components/schemas/PipelineRunSummary'
        sectionErrors:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Diagnostic'
    PipelineSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        sourceAsset:
          type: string
        mode:
          type: string
        ownerName:
          type: string
        recentRunAt:
          type: string
          format: date-time
...
    CredentialSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        provider:
          type: string
        credentialType:
          type: string
        referenceCount:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        rotatedAt:
          type: string
          format: date-time
    PipelineRunSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        flinkJobId:
          type: string
        status:
          type: string
        pipelineName:
          type: string
        sourceAsset:
          type: string
        triggerType:
          type: string
        readCount:
          type: integer
          format: int64
        writtenCount:
          type: integer
          format: int64
        rejectedCount:
          type: integer
          format: int64
        startedAt:
          type: string
          format: date-time
        durationMs:
          type: integer
          format: int64
...
    ResourceId:
      type: string
      description: Caller-defined stable ID used in API paths and references
      example: employee
      pattern: ^[A-Za-z][A-Za-z0-9_-]{0,159}$
...
    AssetField:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        nullable:
          type: boolean
        inferredType:
          type: string
        originalType:
          type: string
        sensitive:
          type: boolean
        primaryKeyCandidate:
          type: boolean
        sampleValue:
          type: string
...
    CredentialInput:
      type: object
      properties:
        name:
          type: string
        mode:
          type: string
        existingSecretRef:
          type: string
          format: uuid
        fileRefs:
          type: object
          additionalProperties:
            type: string
        values:
          type: object
          additionalProperties:
            type: string
...
    Diagnostic:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        stage:
          type: string
        reason:
          type: string
        suggestion:
          type: string
        technicalDetail:
          type: string
        occurredAt:
          type: string
          format: date-time
...
    TestStage:
      type: object
      properties:
        status:
          type: string
        stage:
          type: string
        message:
          type: string
        durationMs:
          type: integer
          format: int64
...
  parameters:
...
    OntologyId:
      name: ontologyId
      in: path
      description: Caller-defined stable Ontology ID
      required: true
      schema:
        $ref: '#/components/schemas/ResourceId'
...